+++ fiddles/node/fiddle-0013-CouchDB/db-cli.js
@@ -1,0 +1,20 @@
+#!/usr/bin/env node --harmony
+
+"use strict";
+
+const request = require('request'),
+    options = {
+        method: process.argv[2] || 'GET',
+        url: 'http://localhost:5984/' + (process.argv[3] || '')
+    };
+
+request(options, function (err, res, body) {
+    console.log(__dirname);
+    if (err) {
+        throw Error(err);
+    } else {
+        console.log(res.statusCode, JSON.parse(body));
+    }
+});
+
+